HELLO EVERYONE
I  am currently working on an electronic speed controller with back emf as feedback to determine the position of the rotor of my brush-less dc motor. i am at the stage where i have to detect back emf of the motor using the zero cross detection method and the analog comparator module of the atmega328p chip. i have configured it well and i think its supposed to be working well. on my simulation in proteus 8.6, it works well and interrupts at the zero crossing point, but when i wire the circuit the same way as in the simulation, it doesn't work.